Manage all authentication through AuthenticationActivity.
In the prior design, AuthenticationActivity was started asynchronously,
showing authenticating/failure status in response to broadcasts.  Meanwhile,
AuthenticationManager/FlowManager could have started a flow and also the
PourInProgressActivity.

This was excessively asynchronous, and caused problems like issue #41, where
the duration between "authenticating" and "authenticated" was so brief that
the (single top) "authenticating" dialog wasn't displayed by the framework
until *after* the PourInProgress activity had finished.

The new design is also much simpler to follow: AuthenticatingActivity starts,
shows the progress bar, and interacts with (synchronous) AuthenticationManager
apis until the auth status is resolved, at which point it starts the flows
and finishes itself.

There is also now room for the AuthActivity to implement more elaborate UIs
during authentication (for instance, prompting for a pin).

Closes issue #41.