Update for build tools and dependencies
This commit *should* make the project buildable with the latest tools
and without needing usb-serial-for-android locally (uses maven). It's mostly
removing the .idea files and the breaking DSL changes when the android plugin went 1.0.
It also only requires the GCM portions of Google Play Services since we
seem to be butting up against the dex method limit when we include it all.